Separate shared RGB processing code from send_tokens.
