+++ src/operations/rgb.rs
@@ -7,6 +7,7 @@ mod import_asset;
 mod issue_asset;
 mod receive_tokens;
 mod send_tokens;
+mod shared;
 mod validate_transaction;
 
 pub use accept_transaction::accept_transfer;
@@ -14,7 +15,7 @@ pub use create_psbt::create_psbt;
 pub use import_asset::{get_asset_by_genesis, get_assets};
 pub use issue_asset::issue_asset;
 pub use receive_tokens::blind_utxo;
-pub use send_tokens::{transfer_asset, ConsignmentDetails};
+pub use send_tokens::transfer_asset;
 pub use validate_transaction::validate_transfer;
 
 pub use rgb_std::Contract;
